requirements.txt
aiogram==2.25.1
binance-futures-connector==3.1.1
python-dotenv==0.21.1
requests==2.28.2
